requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Google Cloud
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0

# Outras
python-dotenv==1.0.1
